requirements/requirements.txt
Adafruit-GPIO==1.0.3
Adafruit-PCA9685==1.0.1
Adafruit-PureIO==0.2.1
autopep8==1.3.3
decorator==4.1.2
flake8==3.5.0
importmagic==0.1.7
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
mccabe==0.6.1
numpy==1.13.3
parso==0.1.0
pexpect==4.3.0
picamera==1.13
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
RPi.GPIO==0.6.3
simplegeneric==0.8.1
six==1.11.0
spidev==3.2
traitlets==4.3.2
wcwidth==0.1.7
yapf==0.20.0
